+++ CHANGELOG.md
@@ -9,17 +9,28 @@ NOTE: As semantic versioning states all 0.y.z releases can contain breaking chan
 We use *breaking :warning:* to mark changes that are not backward compatible (relates only to v0.y.z releases.)
 
 ## Unreleased
+- [#8190](https://github.com/thanos-io/thanos/pull/8190) Fix markdown formatting in CHANGELOG.
+- [#8202](https://github.com/thanos-io/thanos/pull/8202) Receive: Unhide `--tsdb.enable-native-histograms` flag
+- [#8225](https://github.com/thanos-io/thanos/pull/8225) tools: Extend bucket ls options.
+- [#8136](https://github.com/thanos-io/thanos/pull/8279) docs: legacy --store flags + alerts examples legacy metrics
 
 ### Added
 
+- [#8238](https://github.com/thanos-io/thanos/pull/8238) Receive: add shuffle sharding support
+
 ### Changed
+
+- [#8192](https://github.com/thanos-io/thanos/pull/8192) Sidecar: fix default get config timeout
 
 ### Removed
 
 ### Fixed
-- [#8175](https://github.com/thanos-io/thanos/pull/8175) Query: fix endpointset setup
 
-## [v0.38.0 - <in progress>](https://github.com/thanos-io/thanos/tree/release-0.38)
+- [#8211](https://github.com/thanos-io/thanos/pull/8211) Query: fix panic on nested partial response in distributed instant query
+- [#8216](https://github.com/thanos-io/thanos/pull/8216) Query/Receive: fix iter race between `next()` and `stop()` introduced in https://github.com/thanos-io/thanos/pull/7821.
+- [#8212](https://github.com/thanos-io/thanos/pull/8212) Receive: Ensure forward/replication metrics are incremented in err cases
+
+## [v0.38.0](https://github.com/thanos-io/thanos/tree/release-0.38) - 03.04.2025
 
 ### Fixed
 - [#8091](https://github.com/thanos-io/thanos/pull/8091) *: Add POST into allowed CORS methods header
@@ -37,6 +48,8 @@ We use *breaking :warning:* to mark changes that are not backward compatible (re
 - [#8131](https://github.com/thanos-io/thanos/pull/8131) Store Gateway: Optimize regex matchers for .* and .+. #8131
 - [#7808](https://github.com/thanos-io/thanos/pull/7808) Query: Support chain deduplication algorithm.
 - [#8158](https://github.com/thanos-io/thanos/pull/8158) Rule: Add support for query offset.
+- [#8110](https://github.com/thanos-io/thanos/pull/8110) Compact: implement native histogram downsampling.
+- [#7996](https://github.com/thanos-io/thanos/pull/7996) Receive: Add OTLP endpoint.
 
 ### Changed
 
@@ -44,7 +57,7 @@ We use *breaking :warning:* to mark changes that are not backward compatible (re
 - [#7012](https://github.com/thanos-io/thanos/pull/7012) Query: Automatically adjust `max_source_resolution` based on promql query to avoid querying data from higher resolution resulting empty results.
 - [#8118](https://github.com/thanos-io/thanos/pull/8118) Query: Bumped promql-engine
 - [#8135](https://github.com/thanos-io/thanos/pull/8135) Query: respect partial response in distributed engine
-- [#8136](https://github.com/thanos-io/thanos/pull/8136) Query,Ruler: Documentation change legacy flags
+- [#8181](https://github.com/thanos-io/thanos/pull/8181) Deps: bump promql engine
 
 ### Removed
 
@@ -54,6 +67,7 @@ We use *breaking :warning:* to mark changes that are not backward compatible (re
 
 - [#7970](https://github.com/thanos-io/thanos/pull/7970) Sidecar: Respect min-time setting.
 - [#7962](https://github.com/thanos-io/thanos/pull/7962) Store: Fix potential deadlock in hedging request.
+- [#8175](https://github.com/thanos-io/thanos/pull/8175) Query: fix endpointset setup
 
 ### Added